initial commit
I've set up the initial structure... I want to ensure that the "core"
styleguide files are separate from the pattern library style and html,
so need to work on either better documentation or file structure for
that.

Currently, the app runs, and has a simple helper called `pattern` that
will display the pattern file in a iframe by itself, and show it's code
below it using coderay.

The next phase is getting the pattern library css/js into the proxied
iframe, and documenting the style. I'll make a base structure and menu
for the pattern app itself, and set up some test modules.