add beginnings of style and behavior patterns
I have added a pattern specific stylesheet and javascript folder. I'm
not a fan of how confusing it is at present, ideally I would like to
move the "parent app" styling to a folder called `core` or something
like that and have the pattern stuff in the root, but for now this will
do.